requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
requests
python-dotenv
pytest
alembic
asyncpg
redis
aioredis
gunicorn
pymongo
logging